move errors to common, update reducer
